+++ src/api/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.31.2] - 2026-01-15
+
 ### Fixed
 
 - Force `Session.end` naive date times to be aware
@@ -669,7 +671,8 @@ update` command
 
 - Implement base FastAPI app
 
-[unreleased]: https://github.com/MTES-MCT/qualicharge/compare/v0.31.1...main
+[unreleased]: https://github.com/MTES-MCT/qualicharge/compare/v0.31.2...main
+[0.31.2]: https://github.com/MTES-MCT/qualicharge/compare/v0.31.1...v0.31.2
 [0.31.1]: https://github.com/MTES-MCT/qualicharge/compare/v0.31.0...v0.31.1
 [0.31.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.30.1...v0.31.0
 [0.30.1]: https://github.com/MTES-MCT/qualicharge/compare/v0.30.0...v0.30.1